+++ app/abci_prepare_proposal.go
@@ -27,7 +27,7 @@ func (app *App) NewProposalContext(header cmtproto.Header) sdk.Context {
 	return ctx
 }
 
-// PrepareProposal fulfills the celestia-core version of the ABCI interface by
+// PrepareProposal fulfills the sunrise-core version of the ABCI interface by
 // preparing the proposal block data. The square size is determined by first
 // estimating it via the size of the passed block data. Then, this method
 // generates the data root for the proposal block and passes it back to
