+++ app/consts.go
@@ -22,13 +22,17 @@ import (
 	auth "github.com/cosmos/cosmos-sdk/x/auth"
 	"github.com/cosmos/cosmos-sdk/x/auth/vesting"
 	authzmodule "github.com/cosmos/cosmos-sdk/x/authz/module"
+	distribution "github.com/cosmos/cosmos-sdk/x/distribution"
 	"github.com/cosmos/cosmos-sdk/x/genutil"
 	mint "github.com/cosmos/cosmos-sdk/x/mint"
 	"github.com/cosmos/cosmos-sdk/x/params"
+	slashing "github.com/cosmos/cosmos-sdk/x/slashing"
 
 	capability "github.com/cosmos/ibc-go/modules/capability"
 	"github.com/cosmos/ibc-go/v8/modules/apps/transfer"
 	ibc "github.com/cosmos/ibc-go/v8/modules/core"
+
+	"github.com/sunriselayer/sunrise/app/defaultoverrides"
 )
 
 var (
@@ -69,18 +73,20 @@ func ModuleBasics() sdkmodule.BasicManager {
 	config.SetBech32PrefixForValidator(validatorAddressPrefix, validatorPubKeyPrefix)
 	config.SetBech32PrefixForConsensusNode(consNodeAddressPrefix, consNodePubKeyPrefix)
 
-	moduleBasics := sdkmodule.NewBasicManager(
+	moduleBasics := sdkmodule.BasicManager{}
+
+	moduleBasics = sdkmodule.NewBasicManager(
 		auth.AppModuleBasic{},
 		genutil.AppModuleBasic{},
-		bankModule{},
+		defaultoverrides.BankModuleBasic{},
 		capability.AppModuleBasic{},
-		stakingModule{},
+		defaultoverrides.StakingModuleBasic{},
 		mint.AppModuleBasic{},
-		distributionModule{},
-		govModule{},
+		distribution.AppModuleBasic{},
+		defaultoverrides.GovModuleBasic{},
 		params.AppModuleBasic{},
-		crisisModule{},
-		slashingModule{},
+		defaultoverrides.CrisisModuleBasic{},
+		slashing.AppModuleBasic{},
 		authzmodule.AppModuleBasic{},
 		feegrantmodule.AppModuleBasic{},
 		ibc.AppModuleBasic{},
@@ -96,7 +102,7 @@ func ModuleBasics() sdkmodule.BasicManager {
 		feemodule.AppModuleBasic{},
 		upgrade.AppModuleBasic{},
 	)
-	// moduleBasics := sdkmodule.BasicManager{}
+
 	// depinject.Inject(
 	// 	depinject.Configs(AppConfig()),
 	// 	&moduleBasics,
