--- app/default_overrides.go
@@ -1,251 +1,0 @@
-package app
-
-import (
-	"encoding/json"
-	"fmt"
-	"time"
-
-	sdkmath "cosmossdk.io/math"
-	tmcfg "github.com/cometbft/cometbft/config"
-	tmproto "github.com/cometbft/cometbft/proto/tendermint/types"
-	coretypes "github.com/cometbft/cometbft/types"
-	"github.com/cosmos/cosmos-sdk/codec"
-	serverconfig "github.com/cosmos/cosmos-sdk/server/config"
-	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/cosmos/cosmos-sdk/x/bank"
-	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
-	"github.com/cosmos/cosmos-sdk/x/crisis"
-	crisistypes "github.com/cosmos/cosmos-sdk/x/crisis/types"
-	distribution "github.com/cosmos/cosmos-sdk/x/distribution"
-	distributiontypes "github.com/cosmos/cosmos-sdk/x/distribution/types"
-	"github.com/cosmos/cosmos-sdk/x/gov"
-	govtypes "github.com/cosmos/cosmos-sdk/x/gov/types/v1"
-	"github.com/cosmos/cosmos-sdk/x/slashing"
-	slashingtypes "github.com/cosmos/cosmos-sdk/x/slashing/types"
-	"github.com/cosmos/cosmos-sdk/x/staking"
-	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
-	"github.com/sunriselayer/sunrise/pkg/appconsts"
-)
-
-// bankModule defines a custom wrapper around the x/bank module's AppModuleBasic
-// implementation to provide custom default genesis state.
-type bankModule struct {
-	bank.AppModuleBasic
-}
-
-// DefaultGenesis returns custom x/bank module genesis state.
-func (bankModule) DefaultGenesis(cdc codec.JSONCodec) json.RawMessage {
-	metadataFee := banktypes.Metadata{
-		Description: "The native token of the Sunrise network for fees.",
-		DenomUnits: []*banktypes.DenomUnit{
-			{
-				Denom:    "urise",
-				Exponent: 0,
-			},
-			{
-				Denom:    "rise",
-				Exponent: 6,
-			},
-		},
-		Base:    "urise",
-		Display: "rise",
-		Name:    "Sunrise RISE",
-		Symbol:  "RISE",
-	}
-	metadataBond := banktypes.Metadata{
-		Description: "The native token of the Sunrise network for staking. This token is non transferrable. This token can be retrieved by providing liquidity.",
-		DenomUnits: []*banktypes.DenomUnit{
-			{
-				Denom:    "uvrise",
-				Exponent: 0,
-			},
-			{
-				Denom:    "vrise",
-				Exponent: 6,
-			},
-		},
-		Base:    "uvrise",
-		Display: "vrise",
-		Name:    "Sunrise VRISE",
-		Symbol:  "VRISE",
-	}
-
-	sendEnabledVrise := banktypes.SendEnabled{
-		Denom:   "uvrise",
-		Enabled: false,
-	}
-
-	genState := banktypes.DefaultGenesisState()
-	genState.DenomMetadata = append(genState.DenomMetadata, metadataFee)
-	genState.DenomMetadata = append(genState.DenomMetadata, metadataBond)
-
-	genState.SendEnabled = append(genState.SendEnabled, sendEnabledVrise)
-
-	return cdc.MustMarshalJSON(genState)
-}
-
-// stakingModule wraps the x/staking module in order to overwrite specific
-// ModuleManager APIs.
-type stakingModule struct {
-	staking.AppModuleBasic
-}
-
-// DefaultGenesis returns custom x/staking module genesis state.
-func (stakingModule) DefaultGenesis(cdc codec.JSONCodec) json.RawMessage {
-	params := stakingtypes.DefaultParams()
-	params.UnbondingTime = appconsts.DefaultUnbondingTime
-	params.BondDenom = appconsts.BondDenom
-	params.MinCommissionRate = sdkmath.LegacyNewDecWithPrec(5, 2) // 5%
-
-	return cdc.MustMarshalJSON(&stakingtypes.GenesisState{
-		Params: params,
-	})
-}
-
-// stakingModule wraps the x/staking module in order to overwrite specific
-// ModuleManager APIs.
-type slashingModule struct {
-	slashing.AppModuleBasic
-}
-
-// DefaultGenesis returns custom x/staking module genesis state.
-func (slashingModule) DefaultGenesis(cdc codec.JSONCodec) json.RawMessage {
-	params := slashingtypes.DefaultParams()
-	params.MinSignedPerWindow = sdkmath.LegacyNewDecWithPrec(75, 2) // 75%
-	params.SignedBlocksWindow = 5000
-	params.DowntimeJailDuration = time.Minute * 1
-	params.SlashFractionDoubleSign = sdkmath.LegacyNewDecWithPrec(2, 2) // 2%
-	params.SlashFractionDowntime = sdkmath.LegacyZeroDec()              // 0%
-
-	return cdc.MustMarshalJSON(&slashingtypes.GenesisState{
-		Params: params,
-	})
-}
-
-type crisisModule struct {
-	crisis.AppModuleBasic
-}
-
-// DefaultGenesis returns custom x/crisis module genesis state.
-func (crisisModule) DefaultGenesis(cdc codec.JSONCodec) json.RawMessage {
-	return cdc.MustMarshalJSON(&crisistypes.GenesisState{
-		ConstantFee: sdk.NewCoin(appconsts.BondDenom, sdkmath.NewInt(1000)),
-	})
-}
-
-type distributionModule struct {
-	distribution.AppModuleBasic
-}
-
-// DefaultGenesis returns custom x/distribution module genesis state.
-func (distributionModule) DefaultGenesis(cdc codec.JSONCodec) json.RawMessage {
-	params := distributiontypes.DefaultParams()
-	params.BaseProposerReward = sdkmath.LegacyZeroDec()  // 0%
-	params.BonusProposerReward = sdkmath.LegacyZeroDec() // 0%
-	return cdc.MustMarshalJSON(&distributiontypes.GenesisState{
-		Params: params,
-	})
-}
-
-// govModule is a custom wrapper around the x/gov module's AppModuleBasic
-// implementation to provide custom default genesis state.
-type govModule struct {
-	gov.AppModuleBasic
-}
-
-// DefaultGenesis returns custom x/gov module genesis state.
-func (govModule) DefaultGenesis(cdc codec.JSONCodec) json.RawMessage {
-	genState := govtypes.DefaultGenesisState()
-	day := time.Duration(time.Hour * 24)
-	oneWeek := day * 7
-
-	genState.Params.MinDeposit = sdk.NewCoins(sdk.NewCoin(appconsts.BondDenom, sdkmath.NewInt(10_000_000_000))) // 10,000 TIA
-	genState.Params.MaxDepositPeriod = &oneWeek
-	genState.Params.VotingPeriod = &oneWeek
-
-	return cdc.MustMarshalJSON(genState)
-}
-
-// DefaultConsensusParams returns a ConsensusParams with a MaxBytes
-// determined using a goal square size.
-func DefaultConsensusParams() *tmproto.ConsensusParams {
-	block := DefaultBlockParams()
-	evidence := DefaultEvidenceParams()
-	validatorTmp := coretypes.DefaultValidatorParams()
-	validator := tmproto.ValidatorParams{
-		PubKeyTypes: validatorTmp.PubKeyTypes,
-	}
-	return &tmproto.ConsensusParams{
-		Block:     &block,
-		Evidence:  &evidence,
-		Validator: &validator,
-		Version: &tmproto.VersionParams{
-			App: appconsts.LatestVersion,
-		},
-	}
-}
-
-// DefaultBlockParams returns a default BlockParams with a MaxBytes determined
-// using a goal square size.
-func DefaultBlockParams() tmproto.BlockParams {
-	return tmproto.BlockParams{
-		MaxBytes: appconsts.DefaultMaxBytes,
-		MaxGas:   -1,
-	}
-}
-
-// DefaultEvidenceParams returns a default EvidenceParams with a MaxAge
-// determined using a goal block time.
-func DefaultEvidenceParams() tmproto.EvidenceParams {
-	evdParamsTmp := coretypes.DefaultEvidenceParams()
-	evdParams := tmproto.EvidenceParams{
-		MaxBytes: evdParamsTmp.MaxBytes,
-	}
-	evdParams.MaxAgeDuration = appconsts.DefaultUnbondingTime
-	evdParams.MaxAgeNumBlocks = int64(appconsts.DefaultUnbondingTime.Seconds())/int64(appconsts.GoalBlockTime.Seconds()) + 1
-	return evdParams
-}
-
-func DefaultConsensusConfig() *tmcfg.Config {
-	cfg := tmcfg.DefaultConfig()
-	// Set broadcast timeout to be 50 seconds in order to avoid timeouts for long block times
-	// TODO: make TimeoutBroadcastTx configurable per https://github.com/celestiaorg/celestia-app/issues/1034
-	cfg.RPC.TimeoutBroadcastTxCommit = 50 * time.Second
-	cfg.RPC.MaxBodyBytes = int64(8388608) // 8 MiB
-
-	// cfg.Mempool.TTLNumBlocks = 5
-	// cfg.Mempool.TTLDuration = time.Duration(cfg.Mempool.TTLNumBlocks) * appconsts.GoalBlockTime
-	// Given that there is a stateful transaction size check in CheckTx,
-	// We set a loose upper bound on what we expect the transaction to
-	// be based on the upper bound size of the entire block for the given
-	// version. This acts as a first line of DoS protection
-	upperBoundBytes := appconsts.DefaultSquareSizeUpperBound * appconsts.DefaultSquareSizeUpperBound * appconsts.ContinuationSparseShareContentSize
-	cfg.Mempool.MaxTxBytes = upperBoundBytes
-	// cfg.Mempool.MaxTxsBytes = int64(upperBoundBytes) * cfg.Mempool.TTLNumBlocks
-	// cfg.Mempool.Version = "v1" // prioritized mempool
-
-	cfg.Consensus.TimeoutPropose = appconsts.TimeoutPropose
-	cfg.Consensus.TimeoutCommit = appconsts.TimeoutCommit
-	cfg.Consensus.SkipTimeoutCommit = false
-
-	cfg.TxIndex.Indexer = "null"
-	cfg.Storage.DiscardABCIResponses = true
-
-	return cfg
-}
-
-func DefaultAppConfig() *serverconfig.Config {
-	cfg := serverconfig.DefaultConfig()
-	cfg.API.Enable = false
-	cfg.GRPC.Enable = false
-	cfg.GRPCWeb.Enable = false
-
-	// the default snapshot interval was determined by picking a large enough
-	// value as to not dramatically increase resource requirements while also
-	// being greater than zero so that there are more nodes that will serve
-	// snapshots to nodes that state sync
-	cfg.StateSync.SnapshotInterval = 1500
-	cfg.StateSync.SnapshotKeepRecent = 2
-	cfg.MinGasPrices = fmt.Sprintf("%v%s", appconsts.DefaultMinGasPrice, appconsts.BondDenom)
-	return cfg
-}
