+++ app/defaultoverrides/crisis.go
@@ -1,0 +1,23 @@
+package defaultoverrides
+
+import (
+	"encoding/json"
+
+	sdkmath "cosmossdk.io/math"
+	"github.com/cosmos/cosmos-sdk/codec"
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	crisis "github.com/cosmos/cosmos-sdk/x/crisis"
+	crisistypes "github.com/cosmos/cosmos-sdk/x/crisis/types"
+	"github.com/sunriselayer/sunrise/pkg/appconsts"
+)
+
+type CrisisModuleBasic struct {
+	crisis.AppModuleBasic
+}
+
+// DefaultGenesis returns custom x/crisis module genesis state.
+func (CrisisModuleBasic) DefaultGenesis(cdc codec.JSONCodec) json.RawMessage {
+	return cdc.MustMarshalJSON(&crisistypes.GenesisState{
+		ConstantFee: sdk.NewCoin(appconsts.BondDenom, sdkmath.NewInt(1000)),
+	})
+}
