+++ app/defaultoverrides/staking.go
@@ -1,0 +1,27 @@
+package defaultoverrides
+
+import (
+	"encoding/json"
+
+	"github.com/cosmos/cosmos-sdk/codec"
+	staking "github.com/cosmos/cosmos-sdk/x/staking"
+	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
+	"github.com/sunriselayer/sunrise/pkg/appconsts"
+)
+
+// StakingModuleBasic wraps the x/staking module in order to overwrite specific
+// ModuleManager APIs.
+type StakingModuleBasic struct {
+	staking.AppModuleBasic
+}
+
+// DefaultGenesis returns custom x/staking module genesis state.
+func (StakingModuleBasic) DefaultGenesis(cdc codec.JSONCodec) json.RawMessage {
+	params := stakingtypes.DefaultParams()
+	params.UnbondingTime = appconsts.DefaultUnbondingTime
+	params.BondDenom = appconsts.BondDenom
+
+	return cdc.MustMarshalJSON(&stakingtypes.GenesisState{
+		Params: params,
+	})
+}
